+++ src/Symfony/Bundle/SecurityBundle/DataCollector/SecurityDataCollector.php
@@ -22,9 +22,7 @@ use Symfony\Component\Security\Core\Authentication\Token\SwitchUserToken;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManagerInterface;
 use Symfony\Component\Security\Core\Authorization\TraceableAccessDecisionManager;
 use Symfony\Component\Security\Core\Authorization\Voter\TraceableVoter;
-use Symfony\Component\Security\Core\Role\Role;
 use Symfony\Component\Security\Core\Role\RoleHierarchyInterface;
-use Symfony\Component\Security\Core\Role\SwitchUserRole;
 use Symfony\Component\Security\Http\Firewall\SwitchUserListener;
 use Symfony\Component\Security\Http\FirewallMapInterface;
 use Symfony\Component\Security\Http\Logout\LogoutUrlGenerator;
@@ -92,33 +90,15 @@ class SecurityDataCollector extends DataCollector implements LateDataCollectorIn
             ];
         } else {
             $inheritedRoles = [];
-
-            if (method_exists($token, 'getRoleNames')) {
-                $assignedRoles = $token->getRoleNames();
-            } else {
-                $assignedRoles = array_map(function (Role $role) { return $role->getRole(); }, $token->getRoles(false));
-            }
+            $assignedRoles = $token->getRoleNames();
 
             $impersonatorUser = null;
             if ($token instanceof SwitchUserToken) {
                 $impersonatorUser = $token->getOriginalToken()->getUsername();
-            } else {
-                foreach ($token->getRoles(false) as $role) {
-                    if ($role instanceof SwitchUserRole) {
-                        $impersonatorUser = $role->getSource()->getUsername();
-                        break;
-                    }
-                }
             }
 
             if (null !== $this->roleHierarchy) {
-                if (method_exists($this->roleHierarchy, 'getReachableRoleNames')) {
-                    $allRoles = $this->roleHierarchy->getReachableRoleNames($assignedRoles);
-                } else {
-                    $allRoles = array_map(function (Role $role) { return (string) $role; }, $this->roleHierarchy->getReachableRoles($token->getRoles(false)));
-                }
-
-                foreach ($allRoles as $role) {
+                foreach ($this->roleHierarchy->getReachableRoleNames($assignedRoles) as $role) {
                     if (!\in_array($role, $assignedRoles, true)) {
                         $inheritedRoles[] = $role;
                     }