+++ src/Symfony/Component/Security/Core/Authentication/Token/AbstractToken.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\Security\Core\Authentication\Token;
 
-use Symfony\Component\Security\Core\Role\Role;
 use Symfony\Component\Security\Core\User\EquatableInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
@@ -24,7 +23,6 @@ use Symfony\Component\Security\Core\User\UserInterface;
 abstract class AbstractToken implements TokenInterface
 {
     private $user;
-    private $roles = [];
     private $roleNames = [];
     private $authenticated = false;
     private $attributes = [];
@@ -37,32 +35,16 @@ abstract class AbstractToken implements TokenInterface
     public function __construct(array $roles = [])
     {
         foreach ($roles as $role) {
-            if (\is_string($role)) {
-                $role = new Role($role, false);
-            } elseif (!$role instanceof Role) {
-                throw new \InvalidArgumentException(sprintf('$roles must be an array of strings, or Role instances, but got %s.', \gettype($role)));
-            }
-
-            $this->roles[] = $role;
-            $this->roleNames[] = (string) $role;
+            $this->roleNames[] = $role;
         }
-    }
-
-    public function getRoleNames(): array
-    {
-        return $this->roleNames;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function getRoles()
+    public function getRoleNames(): array
     {
-        if (0 === \func_num_args() || func_get_arg(0)) {
-            @trigger_error(sprintf('The %s() method is deprecated since Symfony 4.3. Use the getRoleNames() method instead.', __METHOD__), E_USER_DEPRECATED);
-        }
-
-        return $this->roles;
+        return $this->roleNames;
     }
 
     /**
@@ -158,7 +140,7 @@ abstract class AbstractToken implements TokenInterface
      */
     public function __serialize(): array
     {
-        return [$this->user, $this->authenticated, $this->roles, $this->attributes, $this->roleNames];
+        return [$this->user, $this->authenticated, null, $this->attributes, $this->roleNames];
     }
 
     /**
@@ -198,15 +180,7 @@ abstract class AbstractToken implements TokenInterface
      */
     public function __unserialize(array $data): void
     {
-        [$this->user, $this->authenticated, $this->roles, $this->attributes] = $data;
-
-        // migration path to 4.3+
-        if (null === $this->roleNames = $data[4] ?? null) {
-            $this->roleNames = [];
-            foreach ($this->roles as $role) {
-                $this->roleNames[] = (string) $role;
-            }
-        }
+        [$this->user, $this->authenticated, , $this->attributes, $this->roleNames] = $data;
     }
 
     /**
@@ -291,8 +265,8 @@ abstract class AbstractToken implements TokenInterface
         $class = substr($class, strrpos($class, '\\') + 1);
 
         $roles = [];
-        foreach ($this->roles as $role) {
-            $roles[] = $role->getRole();
+        foreach ($this->roleNames as $role) {
+            $roles[] = $role;
         }
 
         return sprintf('%s(user="%s", authenticated=%s, roles="%s")', $class, $this->getUsername(), json_encode($this->authenticated), implode(', ', $roles));