+++ src/Symfony/Component/Security/Core/Authentication/Token/TokenInterface.php
@@ -11,17 +11,14 @@
 
 namespace Symfony\Component\Security\Core\Authentication\Token;
 
-use Symfony\Component\Security\Core\Role\Role;
-
 /**
  * TokenInterface is the interface for the user authentication information.
  *
  * @author Fabien Potencier <fabien@symfony.com>
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
  *
- * @method array    __serialize()                                                                             Returns all the necessary state of the object for serialization purposes - not implementing it is deprecated since Symfony 4.3
- * @method void     __unserialize(array $data) Restores the object state from an array given by __serialize() - not implementing it is deprecated since Symfony 4.3
- * @method string[] getRoleNames()                                                                            The associated roles - not implementing it is deprecated since Symfony 4.3
+ * @method array __serialize()                                                                             Returns all the necessary state of the object for serialization purposes - not implementing it is deprecated since Symfony 4.3
+ * @method void  __unserialize(array $data) Restores the object state from an array given by __serialize() - not implementing it is deprecated since Symfony 4.3
  */
 interface TokenInterface extends \Serializable
 {
@@ -37,11 +34,9 @@ interface TokenInterface extends \Serializable
     /**
      * Returns the user roles.
      *
-     * @return Role[] An array of Role instances
-     *
-     * @deprecated since Symfony 4.3, use the getRoleNames() method instead
+     * @return string[] The associated roles
      */
-    public function getRoles();
+    public function getRoleNames(): array;
 
     /**
      * Returns the user credentials.